Separating utilities into their own file.
